target/riscv: Deep refactoring of the arbitrary precision code
This commit changes the way the helpers are defined so that dependencies
on the general purpose registers are visible to QEMU IR optimizer, which
avoids buggy code motion.
Did a few optimization along the path by removing all calls to mpfr_init
and memcpy.
Finally, added the support to dynamic rounding mode.