New version of the tests
This new version of tests uses a refactored apmacros.h in which the
order of the registers in the macros follows the original riscv order.
Still some refactoring to do to have AP, double and float versions fused
in such a way that most of the conditional compilation would disapear.
Currently e, newton-raphson and cholesky have been valided in AP.