New version of the mpfr emulation
Main change is a new way to call the mpfr functions inside the helpers,
so as to avoid an mpfr_set if neither the precision nor the exponent
have changed.

Also modifications in the tests and addition of scripts
Setup for measuring the simulation time of e computation.  The
computation itself is now done 10,000 times in e.S to amortize QEMU
translation (in fact tests with and without amortizing could be
interesting).
The e-time.zsh script launches 10 runs for degrees 10 to 100 and
precision 53 to 4000, agregates the results and produces a summary file.
Each run produces its results in the trial-n directory.
The e-sum.zsh script gather the values and generated a python file which
produces a plot (./e-sum.zsh|python -) as a pdf for inclusing in a
document.