Arbitrary Precision: Rebase on 63011373ad
This patch is a rebase of the arbitrary precision support on 63011373ad,
Oct 12, 2023.
Quite a lot of changes since they were changes in tcg and new utility
functions since the support of half floats.
The set of patches will need to be worked on to make it easier to apply,
as it currently has an history of various uninteresting trials.